+++ src/checks/index.ts
@@ -1,2 +1,2 @@
 export { runChecks as runChecksForIOS } from './ios';
-export { runChecks as runChecksForReactNative } from './react_native';
+export { runChecks as runChecksForReactNative } from './reactnative';
